Use JSON for title sequences
