--- data/title/openrct2/script.txt
@@ -1,84 +1,0 @@
-# OpenRCT2 0.0.4 Title Sequence Script
-# Arranged by NWSpacek (Kelson Blakewood).
-# If you have parks you think would work well in a title sequence (must be NCSO), contact me on reddit at u/spacek_toast
-
-
-LOAD Ill Paradise.sv6
-LOCATION 113 105
-WAIT 15
-
-LOCATION 120 102
-ROTATE 3
-WAIT 10
-
-LOCATION 73 136
-ROTATE 1
-WAIT 15
-
-LOAD Title Sequence Park 1.sv6
-ROTATE 2
-LOCATION 120 43
-WAIT 15
-
-LOCATION 124 74
-WAIT 15
-
-ROTATE 1
-LOCATION 120 90
-WAIT 10
-
-LOCATION 123 130
-WAIT 20
-
-LOCATION 70 85
-ROTATE 1
-WAIT 15
-
-LOAD Cypress Grove Title.sv6
-LOCATION 106 43
-ROTATE 3
-WAIT 15
-
-LOCATION 60 103
-ROTATE 1
-WAIT 10
-
-LOCATION 91 83
-WAIT 10
-
-ROTATE 1
-LOCATION 117 61
-WAIT 10
-
-LOCATION 82 95
-ROTATE 1
-WAIT 20
-
-LOCATION 110 94
-ROTATE 3
-WAIT 15
-
-LOCATION 79 94
-WAIT 10
-
-LOCATION 63 97
-ROTATE 1
-WAIT 10
-
-LOAD TaCW.sv6
-WAIT 15
-
-ROTATE 1
-LOCATION 110 95
-WAIT 15
-
-LOCATION 57 101
-ROTATE 3
-WAIT 15
-
-LOCATION 100 63
-ROTATE 2
-WAIT 15
-
-RESTART
-END
