--- data/title/rct1/script.txt
@@ -1,98 +1,0 @@
-# RollerCoaster Tycoon 1 Title Sequence
-
-##############################
-# RollerCoaster Tycoon 1     #
-##############################
-# SC_DIAMOND_HEIGHTS
-LOADRCT1 3
-ROTATE 2
-LOCATION 47, 99
-WAIT 15
-ROTATE 2
-LOCATION 90, 48
-WAIT 15
-ROTATE 1
-LOCATION 86, 43
-WAIT 15
-ROTATE 3
-LOCATION 57, 50
-WAIT 15
-ROTATE 2
-LOCATION 86, 64
-WAIT 15
-
-# SC_KATIES_DREAMLAND
-LOADRCT1 7
-LOCATION 48, 44
-WAIT 15
-LOCATION 64, 41
-WAIT 15
-
-# SC_POKEY_PARK,
-LOADRCT1 8
-ROTATE 2
-LOCATION 56, 64
-WAIT 15
-
-# SC_WHITE_WATER_PARK
-LOADRCT1 9
-ROTATE 3
-LOCATION 96, 88
-WAIT 15
-LOCATION 84, 95
-WAIT 15
-
-# SC_MILLENNIUM_MINES
-LOADRCT1 10
-ROTATE 1
-ROTATE 1
-LOCATION 64, 37
-WAIT 15
-
-# SC_KARTS_COASTERS
-LOADRCT1 11
-LOCATION 84, 56
-WAIT 15
-LOCATION 34, 36
-WAIT 15
-LOCATION 33, 65
-WAIT 15
-
-# SC_MELS_WORLD
-LOADRCT1 12
-ROTATE 3
-LOCATION 50, 50
-WAIT 15
-LOCATION 35, 39
-WAIT 15
-ROTATE 3
-LOCATION 62, 39
-WAIT 15
-
-# SC_CRUMBLY_WOODS
-LOADRCT1 15
-ROTATE 3
-LOCATION 57, 94
-WAIT 15
-LOCATION 44, 84
-WAIT 15
-ROTATE 1
-LOCATION 76, 59
-WAIT 15
-LOCATION 76, 48
-WAIT 15
-
-# SC_LIGHTNING_PEAKS
-LOADRCT1 17
-LOCATION 80, 49
-WAIT 15
-LOCATION 60, 62
-WAIT 15
-
-# SC_IVORY_TOWERS
-LOADRCT1 18
-ROTATE 3
-LOCATION 50, 77
-WAIT 15
-
-RESTART
