--- data/title/rct1aa/script.txt
@@ -1,187 +1,0 @@
-# RollerCoaster Tycoon 1 Title Sequence
-
-##############################
-# Added Attractions          #
-##############################
-# SC_HAUNTED_HARBOR
-LOADRCT1 45
-LOCATION 62, 53
-WAIT 11
-ROTATE 3
-LOCATION 78, 45
-WAIT 16
-
-# SC_CANARY_MINES
-LOADRCT1 42
-LOCATION 52, 28
-WAIT 15
-
-# SC_GENTLE_GLEN
-LOADRCT1 48
-LOCATION 68, 45
-WAIT 11
-
-# SC_FUNTOPIA
-LOADRCT1 44
-LOCATION 69, 93
-WAIT 17
-
-# SC_SPRIGHTLY_PARK
-LOADRCT1 51
-ROTATE 3
-LOCATION 97, 72
-WAIT 12
-ROTATE 2
-LOCATION 74, 71
-WAIT 12
-ROTATE 2
-LOCATION 67, 90
-WAIT 12
-
-# SC_GIGGLE_DOWNS
-LOADRCT1 65
-LOCATION 94, 64
-WAIT 10
-
-# SC_SWAMP_COVE
-LOADRCT1 59
-LOCATION 96, 22
-WAIT 10
-ROTATE 3
-LOCATION 90, 29
-WAIT 10
-
-# SC_FRUIT_FARM
-LOADRCT1 53
-LOCATION 47, 52
-WAIT 12
-
-# SC_FUTURE_WORLD
-LOADRCT1 47
-LOCATION 67, 59
-WAIT 19
-
-# SC_THREE_MONKEYS_PARK
-LOADRCT1 41
-ROTATE 2
-LOCATION 72, 61
-WAIT 8
-ROTATE 2
-LOCATION 68, 70
-WAIT 8
-LOCATION 58, 56
-WAIT 8
-ROTATE 1
-LOCATION 44, 70
-WAIT 8
-ROTATE 3
-LOCATION 43, 90
-WAIT 10
-
-# SC_ROMAN_VILLAGE
-LOADRCT1 58
-LOCATION 49, 46
-WAIT 15
-
-# SC_ADRENALINE_HEIGHTS
-LOADRCT1 60
-ROTATE 3
-LOCATION 38, 57
-WAIT 15
-
-##############################
-# RollerCoaster Tycoon 1     #
-##############################
-# SC_DIAMOND_HEIGHTS
-LOADRCT1 3
-ROTATE 2
-LOCATION 47, 99
-WAIT 15
-ROTATE 2
-LOCATION 90, 48
-WAIT 15
-ROTATE 1
-LOCATION 86, 43
-WAIT 15
-ROTATE 3
-LOCATION 57, 50
-WAIT 15
-ROTATE 2
-LOCATION 86, 64
-WAIT 15
-
-# SC_KATIES_DREAMLAND
-LOADRCT1 7
-LOCATION 48, 44
-WAIT 15
-LOCATION 64, 41
-WAIT 15
-
-# SC_POKEY_PARK,
-LOADRCT1 8
-ROTATE 2
-LOCATION 56, 64
-WAIT 15
-
-# SC_WHITE_WATER_PARK
-LOADRCT1 9
-ROTATE 3
-LOCATION 96, 88
-WAIT 15
-LOCATION 84, 95
-WAIT 15
-
-# SC_MILLENNIUM_MINES
-LOADRCT1 10
-ROTATE 1
-ROTATE 1
-LOCATION 64, 37
-WAIT 15
-
-# SC_KARTS_COASTERS
-LOADRCT1 11
-LOCATION 84, 56
-WAIT 15
-LOCATION 34, 36
-WAIT 15
-LOCATION 33, 65
-WAIT 15
-
-# SC_MELS_WORLD
-LOADRCT1 12
-ROTATE 3
-LOCATION 50, 50
-WAIT 15
-LOCATION 35, 39
-WAIT 15
-ROTATE 3
-LOCATION 62, 39
-WAIT 15
-
-# SC_CRUMBLY_WOODS
-LOADRCT1 15
-ROTATE 3
-LOCATION 57, 94
-WAIT 15
-LOCATION 44, 84
-WAIT 15
-ROTATE 1
-LOCATION 76, 59
-WAIT 15
-LOCATION 76, 48
-WAIT 15
-
-# SC_LIGHTNING_PEAKS
-LOADRCT1 17
-LOCATION 80, 49
-WAIT 15
-LOCATION 60, 62
-WAIT 15
-
-# SC_IVORY_TOWERS
-LOADRCT1 18
-ROTATE 3
-LOCATION 50, 77
-WAIT 15
-
-RESTART
