--- data/title/rct2/script.txt
@@ -1,34 +1,0 @@
-# RollerCoaster Tycoon 2 Title Sequence
-
-LOADMM
-
-LOCATION 210 112
-WAIT 13
-
-ROTATE 1
-LOCATION 210 112
-WAIT 14
-
-ROTATE 3
-LOCATION 167 180
-WAIT 12
-
-ROTATE 1
-LOCATION 155 189
-WAIT 12
-
-LOCATION 106 39
-WAIT 12
-
-LOCATION 182 50
-WAIT 12
-
-ROTATE 3
-LOCATION 209 47
-WAIT 12
-
-ROTATE 1
-LOCATION 159 93
-WAIT 12
-
-RESTART
