+++ openrct2.vcxproj
@@ -92,7 +92,7 @@
     <ClCompile Include="src\interface\graph.c" />
     <ClCompile Include="src\interface\keyboard_shortcut.c" />
     <ClCompile Include="src\interface\screenshot.c" />
-    <ClCompile Include="src\interface\title_sequences.c" />
+    <ClCompile Include="src\interface\title_sequences.cpp" />
     <ClCompile Include="src\interface\viewport.c" />
     <ClCompile Include="src\interface\viewport_interaction.c" />
     <ClCompile Include="src\interface\widget.c" />
