+++ src/config.c
@@ -1105,8 +1105,6 @@ bool config_shortcut_keys_save()
 
 #pragma region Title Sequences
 
-static void title_sequence_open(const char *path, const char *customName);
-
 void title_sequences_set_default()
 {
 	char path[MAX_PATH];
@@ -1191,189 +1189,5 @@ void title_sequences_load_presets()
 	}
 	gCurrentPreviewTitleSequence = gCurrentTitleSequence;
 }
-
-static void title_sequence_open(const char *path, const char *customName)
-{
-	utf8 titlePath[MAX_PATH], scriptPath[MAX_PATH];
-	file_info fileInfo;
-	SDL_RWops *file;
-	int fileEnumHandle, i, preset;
-	char parts[3 * 128], *token, *part1, *part2;
-
-	// Check for the script file
-	safe_strcpy(scriptPath, path, sizeof(scriptPath));
-	safe_strcat_path(scriptPath, "script.txt", sizeof(scriptPath));
-	if (!platform_file_exists(scriptPath)) {
-		// No script file, title sequence is invalid
-		return;
-	}
-
-	// Check if the preset is already loaded
-	// No need to read the first two presets as they're hardcoded in
-	for (preset = 0; preset < gConfigTitleSequences.num_presets; preset++) {
-		if (_stricmp(path, gConfigTitleSequences.presets[preset].name) == 0) {
-			return;
-		}
-	}
-	// Otherwise allocate one
-	if (preset == gConfigTitleSequences.num_presets) {
-		gConfigTitleSequences.num_presets++;
-		gConfigTitleSequences.presets = realloc(gConfigTitleSequences.presets, sizeof(title_sequence) * (size_t)gConfigTitleSequences.num_presets);
-
-		if (customName == NULL) {
-			char nameBuffer[MAX_PATH];
-			safe_strcpy(nameBuffer, path, MAX_PATH);
-			// Get folder name
-			// First strip off the last folder separator
-			*strrchr(nameBuffer, *PATH_SEPARATOR) = '\0';
-			// Then find the name of the folder
-			char *name = strrchr(nameBuffer, *PATH_SEPARATOR) + 1;
-			safe_strcpy(gConfigTitleSequences.presets[preset].name, name, TITLE_SEQUENCE_NAME_SIZE);
-			gConfigTitleSequences.presets[preset].path[0] = 0;
-		}
-		else {
-			safe_strcpy(gConfigTitleSequences.presets[preset].name, customName, TITLE_SEQUENCE_NAME_SIZE);
-			safe_strcpy(gConfigTitleSequences.presets[preset].path, path, MAX_PATH);
-		}
-
-		gConfigTitleSequences.presets[preset].saves = malloc(0);
-		gConfigTitleSequences.presets[preset].commands = malloc(0);
-		gConfigTitleSequences.presets[preset].num_saves = 0;
-		gConfigTitleSequences.presets[preset].num_commands = 0;
-	}
-
-	// Get the save file list
-	safe_strcpy(titlePath, path, sizeof(titlePath));
-	safe_strcat_path(titlePath, "*.sv6", sizeof(titlePath));
-	fileEnumHandle = platform_enumerate_files_begin(titlePath);
-	while (platform_enumerate_files_next(fileEnumHandle, &fileInfo)) {
-		gConfigTitleSequences.presets[preset].num_saves++;
-		gConfigTitleSequences.presets[preset].saves = realloc(gConfigTitleSequences.presets[preset].saves, sizeof(char[TITLE_SEQUENCE_MAX_SAVE_LENGTH]) * (size_t)gConfigTitleSequences.presets[preset].num_saves);
-		safe_strcpy(gConfigTitleSequences.presets[preset].saves[gConfigTitleSequences.presets[preset].num_saves - 1], fileInfo.path, TITLE_SEQUENCE_MAX_SAVE_LENGTH);
-		gConfigTitleSequences.presets[preset].saves[gConfigTitleSequences.presets[preset].num_saves - 1][TITLE_SEQUENCE_MAX_SAVE_LENGTH - 1] = '\0';
-	}
-	platform_enumerate_files_end(fileEnumHandle);
-	safe_strcpy(titlePath, path, sizeof(titlePath));
-	safe_strcat_path(titlePath, "*.sc6", sizeof(titlePath));
-	fileEnumHandle = platform_enumerate_files_begin(titlePath);
-	while (platform_enumerate_files_next(fileEnumHandle, &fileInfo)) {
-		gConfigTitleSequences.presets[preset].num_saves++;
-		gConfigTitleSequences.presets[preset].saves = realloc(gConfigTitleSequences.presets[preset].saves, sizeof(char[TITLE_SEQUENCE_MAX_SAVE_LENGTH]) * (size_t)gConfigTitleSequences.presets[preset].num_saves);
-		safe_strcpy(gConfigTitleSequences.presets[preset].saves[gConfigTitleSequences.presets[preset].num_saves - 1], fileInfo.path, TITLE_SEQUENCE_MAX_SAVE_LENGTH);
-		gConfigTitleSequences.presets[preset].saves[gConfigTitleSequences.presets[preset].num_saves - 1][TITLE_SEQUENCE_MAX_SAVE_LENGTH - 1] = '\0';
-	}
-	platform_enumerate_files_end(fileEnumHandle);
-
-	// Load the script file
-	file = SDL_RWFromFile(scriptPath, "r");
-	sint64 fileSize = SDL_RWsize(file);
-	do {
-		title_script_get_line(file, parts);
-
-		token = &parts[0 * 128];
-		part1 = &parts[1 * 128];
-		part2 = &parts[2 * 128];
-		title_command command = { 0 };
-		command.command = 0xFF;
-
-		if (token[0] != 0) {
-			if (_stricmp(token, "LOAD") == 0) {
-				command.command = TITLE_SCRIPT_LOAD;
-				command.saveIndex = 0xFF;
-				for (i = 0; i < gConfigTitleSequences.presets[preset].num_saves && command.saveIndex == 0xFF; i++) {
-					if (_stricmp(part1, gConfigTitleSequences.presets[preset].saves[i]) == 0)
-						command.saveIndex = i;
-				}
-			} else if (_stricmp(token, "LOCATION") == 0) {
-				command.command = TITLE_SCRIPT_LOCATION;
-				command.x = atoi(part1) & 0xFF;
-				command.y = atoi(part2) & 0xFF;
-			} else if (_stricmp(token, "ROTATE") == 0) {
-				command.command = TITLE_SCRIPT_ROTATE;
-				command.rotations = atoi(part1) & 0xFF;
-			} else if (_stricmp(token, "ZOOM") == 0) {
-				command.command = TITLE_SCRIPT_ZOOM;
-				command.zoom = atoi(part1) & 0xFF;
-			} else if (_stricmp(token, "SPEED") == 0) {
-				command.command = TITLE_SCRIPT_SPEED;
-				command.speed = max(1, min(4, atoi(part1) & 0xFF));
-			} else if (_stricmp(token, "WAIT") == 0) {
-				command.command = TITLE_SCRIPT_WAIT;
-				command.seconds = atoi(part1) & 0xFF;
-			} else if (_stricmp(token, "RESTART") == 0) {
-				command.command = TITLE_SCRIPT_RESTART;
-			} else if (_stricmp(token, "END") == 0) {
-				command.command = TITLE_SCRIPT_END;
-			} else if (_stricmp(token, "LOADMM") == 0) {
-				command.command = TITLE_SCRIPT_LOADMM;
-			} else if (_stricmp(token, "LOADRCT1") == 0) {
-				command.command = TITLE_SCRIPT_LOADRCT1;
-				command.saveIndex = atoi(part1) & 0xFF;
-			}
-		}
-		if (command.command != 0xFF) {
-			gConfigTitleSequences.presets[preset].num_commands++;
-			gConfigTitleSequences.presets[preset].commands = realloc(gConfigTitleSequences.presets[preset].commands, sizeof(title_command) * (size_t)gConfigTitleSequences.presets[preset].num_commands);
-			gConfigTitleSequences.presets[preset].commands[gConfigTitleSequences.presets[preset].num_commands - 1] = command;
-		}
-	} while (SDL_RWtell(file) < fileSize);
-	SDL_RWclose(file);
-}
-
-void title_sequence_save_preset_script(int preset)
-{
-	utf8 path[MAX_PATH];
-	SDL_RWops *file;
-	int i;
-
-
-	platform_get_user_directory(path, "title sequences", sizeof(path));
-	safe_strcat_path(path, gConfigTitleSequences.presets[preset].name, MAX_PATH);
-	safe_strcat_path(path, "script.txt", MAX_PATH);
-
-	file = SDL_RWFromFile(path, "wb");
-	if (file == NULL) {
-		log_error("Unable to write to script file.");
-		return;
-	}
-
-	for (i = 0; i < gConfigTitleSequences.presets[preset].num_commands; i++) {
-		title_command *command = &gConfigTitleSequences.presets[preset].commands[i];
-		switch (command->command) {
-		case TITLE_SCRIPT_LOAD:
-			if (command->saveIndex == 0xFF)
-				rwopsprintf(file, "LOAD <No save file>");
-			else
-				rwopsprintf(file, "LOAD %s", gConfigTitleSequences.presets[preset].saves[command->saveIndex]);
-			break;
-		case TITLE_SCRIPT_LOCATION:
-			rwopsprintf(file, "LOCATION %i %i", command->x, command->y);
-			break;
-		case TITLE_SCRIPT_ROTATE:
-			rwopsprintf(file, "ROTATE %i", command->rotations);
-			break;
-		case TITLE_SCRIPT_ZOOM:
-			rwopsprintf(file, "ZOOM %i", command->zoom);
-			break;
-		case TITLE_SCRIPT_SPEED:
-			rwopsprintf(file, "SPEED %i", command->speed);
-			break;
-		case TITLE_SCRIPT_WAIT:
-			rwopsprintf(file, "WAIT %i", command->seconds);
-			rwopswritenewline(file);
-			break;
-		case TITLE_SCRIPT_RESTART:
-			rwopsprintf(file, "RESTART");
-			break;
-		case TITLE_SCRIPT_END:
-			rwopsprintf(file, "END");
-			break;
-		}
-		rwopswritenewline(file);
-	}
-
-	SDL_RWclose(file);
-}
-
 
 #pragma endregion
