+++ src/windows/title_editor.c
@@ -119,8 +119,8 @@ enum WINDOW_TITLE_EDITOR_WIDGET_IDX {
 	//WIDX_TITLE_EDITOR_RELOAD,
 	WIDX_TITLE_EDITOR_SKIP_TO,
 
-	WIDX_TITLE_EDITOR_MOVE_UP,
 	WIDX_TITLE_EDITOR_MOVE_DOWN,
+	WIDX_TITLE_EDITOR_MOVE_UP,
 
 	WIDX_TITLE_EDITOR_REPLAY,
 	WIDX_TITLE_EDITOR_STOP,
