+++ src/world/park.h
@@ -108,8 +108,6 @@ void game_command_remove_park_entrance(int *eax, int *ebx, int *ecx, int *edx, i
 void game_command_set_park_name(int *eax, int *ebx, int *ecx, int *edx, int *esi, int *edi, int *ebp);
 void game_command_buy_land_rights(int *eax, int *ebx, int *ecx, int *edx, int *esi, int *edi, int *ebp);
 
-void map_invalidate_tile(int x, int y, int z0, int z1);
-
 void park_remove_ghost_entrance();
 money32 park_place_ghost_entrance(int x, int y, int z, int direction);
